Add documentation
This adds documentation for the API. The documentation uses Doxygen to
generate the C++ API documentation in XML format, and then uses exhale
to convert the XML to breathe, and then uses the breathe output to
generate sphinx documentation using the ReadTheDocs theme, which
provides a clean, modern, interface for the Doxygen API documentation.

The documentation is simple to build. Simply running make clean && make
html from the docs/ directory will build the documentation at
docs/_build/html/.