Rename and refactor tests
This renames the libbrary from Streamline to Ripple.

Additionally, it refactors the test structure to have the same structure
as the source code, so that test files can be written for specific
implementation files. It also adds the functionality for a global test
suite so that all tests can be run as a single executable, while
keepping the option to run the individual test modules as well.