Add offset_to() for multidim offset calculation
This adds offset_to_soa() and offset_to_aos() functions to compute the
offset to a given index over a given multidimensional space for AoS and
SoA storage, with a specified width for the array type.

This intended use is for any class needing to allocate a large block of
memory for a multidimensional space, and index into that space easily
when the parameters of the space and the intended layout of the data in
the space are known.

Additionally, it is intended to be used by iterator classes over then
space, which need to move around the space.