Add std::disjunction & related type traits
This adds an implementation for std::disjunction to perform a compile
time logical OR of a number of traits. This can be used to determine if
any types in a pack meet a certain trait.

It also adds any_same_v<> to determine if any of the types are the same as
the first type, and any_arithmetic_v<> to determine if any types are
arithmetic.