Add a StridedStorage<Ts...> class
This adds the StridedStorage<Ts...> class which can be used as an
underlying storage type which has strided access over the elements it
stores.

The intention for the class is that it allow classes to store data in an
SoA format (i.e strided) for the performance benefit on the GPU and for
AVX operations. The storage also has an allocator, which can be used to
allocate the memory for the strided storage for given spaces.