Adds the ContiguousStorage<Ts...> class
This adds a ContiguousStorage<Ts...> class for creating and access data
which will be stored contiguously for the Ts types.

The class has the same interface as the StridedStorage<Ts...> class, and
can be used interchangably to allocate and offset into data which is
either strided (SoA) or contiguous (AoS).

The storage does not own the data, but provides the functionality to
determine how much data to allocate for a given space, as well as for
creating the storage type which can then be offset using the allocator.