Add OwnedStorage<Ts...> class
This adds the OwnedStorage<Ts...> class which can store the types Ts...
contiguously in a buffer which it owns.

This can be used by containers when returning a copy of the type which
uses StridedStorage or ContiguousStorage so that the type in local
memory (on the cpu or gpu) will be put into registers and access will be
quicker.