Add layout description and traits
This adds the StorageDescriptor class and AutoLayout interface class,
which can be used by any class for which it would be useful for the
storage type of the class to be deduced at compile time by the container
which stores it, so the the layout is optimal for the intended
architecture. For example, in a class which will be used for processing
on the GPU or using AVX, SoA layout would be more performant. This
functionality allows the class to have this property, without having to
perform the layout or allocation.

The LayoutTraits class is also added, which can be used by allocating
containers to determine how to allocate and layout data, abd which
types to return for different layout types.