Add HostBlock and BlockIterator
This adds the HostBlock class which creates a block of N dimensional
data for some type, which can be accessed through the BlockIterator.

The BlockIterator, in its current form, is used to access an element of
the Block with a pointer-like syntax, so that the Block can store and
access both regular types, and types which implement the StridableLayout
interface so that processing of block data can be performed more
efficiently on such types.

This initial interface will be expended to allow the BlockIterator to
access data in the block relative to the data the iterator points to
(i.e to be able to move around the space for which the block holds
data).