Add offset and shift to BlockIterator
This adds functionality for offsetting and shifting an iterator around a
block.

It also refactors the layout_traits for a type so that the iterator only
stores the underlying storage for the type, and not the type itself, so
that the appropriate type can be returned based on the way that the
iterator is accessed -- so that the iterator behaves like a pointer.

This modified the MultidimSpace interface to remove the option to
compute the step in the space for a given width, but rather as if the
space stores normal types. When this is not the case (i.e when the
layout is known to be different), the step can be modified as
appropriate.

It also adds a pointer wrapper, so that pointer like syntax can be used
on the iterator, regardless of if the iterator stores a pointer, or one
of the custom storage types.