Allow host <-> device block creation
This allows host and device blocks to be created from each other,
copyign the memory and block dimension sizes in between, as appropriate.

It also makes changes to the storage classes to allows the get<> access
method to be used on the device, as well as adding traits for enables
based on dimensionality.