Refactor ExecParams into static and dynamic
This refactors the ExecParams class into an interface for execution
parameters, and adds static and dynamic implementations of the interface
so that the parameters can be given as compile time values or as dynamic
sizes determined at runtime (so that benchmarking can be performed with
the dynamic sizes, and compile time sizes used in release builds).

This also updates the invoke functionality to use the new interface, and
the default sizes for the different dimensions. It also adds tests for
the execution functionality.