Refactor invoke functionality
This completely refactors funciontality for invoking a callable on a
block of data, as well as the classes used to enable the functionality.

ExecParams was refactored to remove the grain size functionlity.
Benchamrking showed no case where grain size was beneficial, and there
was an api cost for having to pass around the parameters, so it was
removed.

Thread, block, and grid indexing was moved to the execution module, as
it makes more sense for it to be there.

Static and dynamic shared memory functionality was added to the invoke
functionality, so that shared memory can be created with both runtime
and compile time sizing.

Storage classes were also updated to allow explicit copying of their
data which can be used in interfacing code to overload copy assignment
and construction from the storage types.