Add working internal boundary loading small 2D
This adds a working implementation for internal iterator boundary
loading for N dimensions. This is tested for an edge case for a small
grid size in 2D. For larger grids, additional functionality needs to be
added to facilitate the test, which will be done next, along with the 3D
tests.