Add execution param support for CPU invoke
This adds support to specify subblock sizes (tiles) to execute over when
using the PCU version of invoke. This is added so that it will later be
possible to invoke a callable on subblocks which are sized depending on
the size of one of the caches, which might improve performance.