Modify reduction predicate to take iterators
This is done for the cases that the data to iterate over is strided,
large, only part of the data type must be modified in the reduction.
Using an iterator gives more flexibility, since in the worst case,
dereferenciong it gives the same result as taking a reference, but for
more complex data types being iterated over, the reduction predicate can
use the data in an way.

Additionally, when the iterator is over strided data, dereferenciong
returns a wrapper type, to which the reference would be, rather than the
underlying type which needs to be modified by the predicate. Using an
iterator solves these problems, and extends the functionality of the
reduction.