Fix iterator creation in execution params
Previously, the iterator creation execution paramters with shared memory
specified created a 3D iteartor in all cases. This causes offsetting
problems for systems which are less than 3D. This commit changes the
creation to require the number of dimensions to be specified so that an
iterator of the correct dimension can be created.

This is not optimal for the dynamic execution parameter case, because it
requires resizing the space held by the parameters. This should be
improved by making the number of dimension for the parameters a template
parameter of the class.