Update Grid pipeline invocation
This adds an overload to grid pipeline invocation which takes additional
grids (i.e in addition to the grid on which the pipeline is invoked)
which allows additional data to be passed to pipeline to be used for the
computation.

It also renames the invoke imlpementation for pipeline to
invoke_pipeline so that the call is more explicit.
]