Refactor Grid pipeline functionality
Thi modifies the way pipelines are applied to grids, to have shared and
non shared versions. Additionally, it allows variadic arguments to
be passed when applying a pipeline, and will offset iterators to any
of the veriadic arguments if the type of the argument is a Grid.

It also changes the defualt implementation to use dynamic shared memory
which the grid, and it's blocks, have padding, and to use statically
allocated shared memory when the grid does not have any padding.