Add Flux interface and implementation
This adds the Flux<> interface which allows the flux to be computed
at the cell between two faces in a given dimension. It also adds
implementations of the interface for the Lax-Friedrichs, Richtmyer, and
Force fluxes.

It only adds a 1D test for the Richtmyer flux, more tests are to be
added for the others and for the other dimensions.