Add iter differencing tests for Stridable types
This adds tests for the differencing operations on a iterator when the
type for the iterator is Strdable. There was a problem with the storage
for owned and view types which caused memory to be accessed for a view
type which had not been allocated. This updates the storage types to fix
this problem.

It also updates the VecImpl type to implement the Stridable interface
and adds overloads for conversions between the different storage
storages types for the vector.